fix signed/unsigned comparison with cast
darcs-hash:20070213124604-c9803-416c7ae383c10919875dfdf3171522e8879e082b.gz

+++ nice/ice-test-client.c
@@ -84,7 +84,7 @@ handle_connection (guint sock)
   line = g_strdup_printf ("H/127.0.0.1/%d/lala/titi\n",
       ntohs (udpsock.addr.port));
 
-  if (write (sock, line, strlen (line)) != strlen (line))
+  if (write (sock, line, strlen (line)) != (gint) strlen (line))
     g_assert_not_reached ();
 
   g_free (line);
